pyproject.toml
[tool.poetry]
name = "getcurrentusd"
version = "0.1.0"
description = ""
authors = ["dimapbbb"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
